+++ azure/storage/file/fileservice.py
@@ -41,6 +41,7 @@ from .._deserialization import (
     _convert_xml_to_service_properties,
     _convert_xml_to_signed_identifiers,
 )
+from ..models import Services
 from .models import FileProperties
 from .._http import HTTPRequest
 from ._chunking import (
@@ -70,7 +71,6 @@ from ._deserialization import (
 from .._deserialization import _parse_properties
 from ..sharedaccesssignature import (
     SharedAccessSignature,
-    ResourceType,
 )
 from ..storageclient import _StorageClient
 from os import path
@@ -189,9 +189,54 @@ class FileService(_StorageClient):
 
         return url
 
-    def generate_shared_access_signature(self, share_name, 
-                                         directory_name=None, 
-                                         file_name=None,
+    def generate_account_shared_access_signature(self, resource_types, permission, 
+                                        expiry, start=None, ip=None, protocol=None):
+        '''
+        Generates a shared access signature for the file service.
+        Use the returned signature with the sas_token parameter of the FileService.
+
+        :param ResourceTypes resource_types:
+            Specifies the resource types that are accessible with the account SAS.
+        :param AccountPermissions permission:
+            The permissions associated with the shared access signature. The 
+            user is restricted to operations allowed by the permissions. 
+            Required unless an id is given referencing a stored access policy 
+            which contains this field. This field must be omitted if it has been 
+            specified in an associated stored access policy.
+        :param expiry:
+            The time at which the shared access signature becomes invalid. 
+            Required unless an id is given referencing a stored access policy 
+            which contains this field. This field must be omitted if it has 
+            been specified in an associated stored access policy. Azure will always 
+            convert values to UTC. If a date is passed in without timezone info, it 
+            is assumed to be UTC.
+        :type expiry: date or str
+        :param start:
+            The time at which the shared access signature becomes valid. If 
+            omitted, start time for this call is assumed to be the time when the 
+            storage service receives the request. Azure will always convert values 
+            to UTC. If a date is passed in without timezone info, it is assumed to 
+            be UTC.
+        :type start: date or str
+        :param str ip:
+            Specifies an IP address or a range of IP addresses from which to accept requests.
+            If the IP address from which the request originates does not match the IP address
+            or address range specified on the SAS token, the request is not authenticated.
+            For example, specifying sip=168.1.5.65 or sip=168.1.5.60-168.1.5.70 on the SAS
+            restricts the request to those IP addresses.
+        :param str protocol:
+            Specifies the protocol permitted for a request made. Possible values are
+            both HTTPS and HTTP (https,http) or HTTPS only (https). The default value
+            is https,http. Note that HTTP only is not a permitted value.
+        '''
+        _validate_not_none('self.account_name', self.account_name)
+        _validate_not_none('self.account_key', self.account_key)
+
+        sas = SharedAccessSignature(self.account_name, self.account_key)
+        return sas.generate_account(Services.FILE, resource_types, permission, 
+                                    expiry, start=start, ip=ip, protocol=protocol)
+
+    def generate_share_shared_access_signature(self, share_name, 
                                          permission=None, 
                                          expiry=None,
                                          start=None, 
@@ -204,25 +249,18 @@ class FileService(_StorageClient):
                                          content_language=None,
                                          content_type=None):
         '''
-        Generates a shared access signature for the share or file.
+        Generates a shared access signature for the share.
         Use the returned signature with the sas_token parameter of FileService.
 
         :param str share_name:
             Name of share.
-        :param str directory_name:
-            Name of directory. SAS tokens cannot be created for directories, so 
-            this parameter should only be present if file_name is provided.
-        :param str file_name:
-            Name of file.
-        :param str permission:
+        :param SharePermissions permission:
             The permissions associated with the shared access signature. The 
             user is restricted to operations allowed by the permissions.
             Permissions must be ordered read, create, write, delete, list.
             Required unless an id is given referencing a stored access policy 
             which contains this field. This field must be omitted if it has been 
             specified in an associated stored access policy.
-            See :class:`.ShareSharedAccessPermissions` and 
-            :class:`.FileSharedAccessPermissions`
         :param expiry:
             The time at which the shared access signature becomes invalid. 
             Required unless an id is given referencing a stored access policy 
@@ -272,32 +310,120 @@ class FileService(_StorageClient):
         _validate_not_none('self.account_name', self.account_name)
         _validate_not_none('self.account_key', self.account_key)
 
-        if file_name:
-            resource_type = ResourceType.RESOURCE_FILE
-            resource_path = share_name
-            if directory_name is not None:
-                resource_path += '/' + _str(directory_name)
-            resource_path += '/' + _str(file_name)
-        else:
-            resource_type = ResourceType.RESOURCE_SHARE
-            resource_path = share_name
-
         sas = SharedAccessSignature(self.account_name, self.account_key)
-        return sas.generate_signed_query_string(
-            'file',
-            resource_path,
-            resource_type,
+        return sas.generate_share(
+            share_name,
             permission, 
             expiry,
-            start, 
-            id,
-            ip,
-            protocol,
-            cache_control,
-            content_disposition,
-            content_encoding,
-            content_language,
-            content_type,
+            start=start, 
+            id=id,
+            ip=ip,
+            protocol=protocol,
+            cache_control=cache_control,
+            content_disposition=content_disposition,
+            content_encoding=content_encoding,
+            content_language=content_language,
+            content_type=content_type,
+        )
+
+    def generate_file_shared_access_signature(self, share_name, 
+                                         directory_name=None, 
+                                         file_name=None,
+                                         permission=None, 
+                                         expiry=None,
+                                         start=None, 
+                                         id=None,
+                                         ip=None,
+                                         protocol=None,
+                                         cache_control=None,
+                                         content_disposition=None,
+                                         content_encoding=None,
+                                         content_language=None,
+                                         content_type=None):
+        '''
+        Generates a shared access signature for the file.
+        Use the returned signature with the sas_token parameter of FileService.
+
+        :param str share_name:
+            Name of share.
+        :param str directory_name:
+            Name of directory. SAS tokens cannot be created for directories, so 
+            this parameter should only be present if file_name is provided.
+        :param str file_name:
+            Name of file.
+        :param FilePermissions permission:
+            The permissions associated with the shared access signature. The 
+            user is restricted to operations allowed by the permissions.
+            Permissions must be ordered read, create, write, delete, list.
+            Required unless an id is given referencing a stored access policy 
+            which contains this field. This field must be omitted if it has been 
+            specified in an associated stored access policy.
+        :param expiry:
+            The time at which the shared access signature becomes invalid. 
+            Required unless an id is given referencing a stored access policy 
+            which contains this field. This field must be omitted if it has 
+            been specified in an associated stored access policy. Azure will always 
+            convert values to UTC. If a date is passed in without timezone info, it 
+            is assumed to be UTC.
+        :type expiry: date or str
+        :param start:
+            The time at which the shared access signature becomes valid. If 
+            omitted, start time for this call is assumed to be the time when the 
+            storage service receives the request. Azure will always convert values 
+            to UTC. If a date is passed in without timezone info, it is assumed to 
+            be UTC.
+        :type start: date or str
+        :param str id:
+            A unique value up to 64 characters in length that correlates to a 
+            stored access policy. To create a stored access policy, use 
+            set_file_service_properties.
+        :param str ip:
+            Specifies an IP address or a range of IP addresses from which to accept requests.
+            If the IP address from which the request originates does not match the IP address
+            or address range specified on the SAS token, the request is not authenticated.
+            For example, specifying sip=168.1.5.65 or sip=168.1.5.60-168.1.5.70 on the SAS
+            restricts the request to those IP addresses.
+        :param str protocol:
+            Specifies the protocol permitted for a request made. Possible values are
+            both HTTPS and HTTP (https,http) or HTTPS only (https). The default value
+            is https,http. Note that HTTP only is not a permitted value.
+        :param str cache_control:
+            Response header value for Cache-Control when resource is accessed
+            using this shared access signature.
+        :param str content_disposition:
+            Response header value for Content-Disposition when resource is accessed
+            using this shared access signature.
+        :param str content_encoding:
+            Response header value for Content-Encoding when resource is accessed
+            using this shared access signature.
+        :param str content_language:
+            Response header value for Content-Language when resource is accessed
+            using this shared access signature.
+        :param str content_type:
+            Response header value for Content-Type when resource is accessed
+            using this shared access signature.
+        '''
+        _validate_not_none('share_name', share_name)
+        _validate_not_none('file_name', file_name)
+        _validate_not_none('self.account_name', self.account_name)
+        _validate_not_none('self.account_key', self.account_key)
+
+        sas = SharedAccessSignature(self.account_name, self.account_key)
+        return sas.generate_file(
+            share_name,
+            directory_name,
+            file_name,
+            permission, 
+            expiry,
+            start=start, 
+            id=id,
+            ip=ip,
+            protocol=protocol,
+            cache_control=cache_control,
+            content_disposition=content_disposition,
+            content_encoding=content_encoding,
+            content_language=content_language,
+            content_type=content_type,
         )
 
     def set_file_service_properties(self, hour_metrics=None, minute_metrics=None,